[Refactor]#74 API Restful 수정

+++ main/src/main/java/com/coggiri/main/commons/jwtUtils/JwtTokenProvider.java
@@ -66,7 +66,7 @@ public class JwtTokenProvider {
     public Authentication getAuthentication(String accessToken){
         Claims claims = parseClaims(accessToken);
         String userName = claims.getSubject();
-        Integer userId = claims.get("userId",Integer.class);
+        Long userId = claims.get("userId",Long.class);
 
         List<UserGroupRole> userRoles = userRepository.findGroupRolesByUserId(userId);
